requirements.txt
fastapi
uvicorn[standard]
pandas
requests
python-dotenv
pymongo
